Unified unit

+++ contracts/OraclePriceFeeds.sol
@@ -29,7 +29,7 @@ contract OraclePriceFeeds {
     
     //srcToken contains weth
     //Dec:8
-    function getPriceTokenToUsdt(address srcTokenAddress) public view returns(uint256,int256){
+    function getPriceTokenToUsdt(address srcTokenAddress) public view returns(uint256,uint256){
         address feedAddress = feedsManager[srcTokenAddress];
         if(feedAddress == address(0)){
             return(0,0);
@@ -42,6 +42,6 @@ contract OraclePriceFeeds {
             ,
             
         )   = priceFeed.latestRoundData();
-        return (10,answer);
+        return (10,uint256(answer));
     }
 }
